patch-id: pass hash-algo to flush_current_id()
After computing patch IDs for a commit, a call to  oid_to_hex() is
used to turn the resulting patch ID to hexadecimal.  If we are using
a hash algorithm different from the_hash_algo, the length of the
hash may be different.

Note that the second "hash" on the output line is the name of the
commit, which is expressed by the_hash_algo; we should continue to
use oid_to_hex() to stringify it.

Signed-off-by: Junio C Hamano <gitster@pobox.com>